requirements.txt
Django==5.2.4
djangorestframework==3.16.0
python-decouple==3.8
Pillow==11.2.1
gunicorn==21.2.0
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
django-cors-headers==4.3.1
django-filter==23.5
whitenoise==6.6.0
dj-database-url==2.1.0
django-extensions==3.2.3
django-debug-toolbar==4.2.0
coverage==7.3.2
factory-boy==3.3.0
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
flake8==6.1.0
black==23.11.0
isort==5.12.0
